+++ apps/heureka-next/src/components/ErrorBoundary/index.ts
@@ -3,4 +3,4 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export { default } from "./ErrorBoundary"
+export { default as ErrorBoundary } from "./ErrorBoundary"
