+++ apps/heureka-next/src/components/Images/index.ts
@@ -3,4 +3,4 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export { default } from "./Images"
+export { default as Images } from "./Images"
