+++ apps/heureka-next/src/components/Navigation/index.ts
@@ -3,4 +3,4 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export { default } from "./Navigation"
+export { default as Navigation } from "./Navigation"
