+++ apps/heureka-next/src/components/Services/index.ts
@@ -3,4 +3,4 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export { default } from "./Services"
+export { default as Services } from "./Services"
