+++ apps/heureka-next/src/components/Shell/ShellContent/index.ts
@@ -3,4 +3,4 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export { default } from "./View"
+export { default as ShellContent } from "./ShellContent"
