+++ apps/heureka-next/src/components/Shell/index.ts
@@ -3,4 +3,4 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export { default } from "./Shell"
+export { default as Shell } from "./Shell"
