+++ apps/heureka-next/src/components/Vulnerabilities/index.ts
@@ -3,4 +3,4 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export { default } from "./Vulnerabilities"
+export { default as Vulnerabilities } from "./Vulnerabilities"
